requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0
webdriver-manager>=3.8.0
PyGithub>=1.57.0
html2text>=2020.1.16
trafilatura>=1.6.0
playwright>=1.30.0
